perf(accessibility): expose viewport contents instead of full scrollback
AXValue now reads GHOSTTY_POINT_VIEWPORT (the bounded visible area)
instead of the whole screen, mirroring upstream's unused
cachedVisibleContents: per-poll reads and accessibilityLine(for:) walks
stay cheap with a large scrollback, and history contents (old tokens)
are no longer handed to any AX client. Selections spanning into
scrollback or alt-drag rectangles degrade to NSNotFound, as documented.

Also verify and pin the uniqueRange substring assumption: read_selection
and read_text share dumpTextLocked -> selectionString with
.unwrap/.trim=false, so linear selections (including soft-wrap
junctions and trailing-blank-trimmed rows) map verbatim — regression
tests added for both shapes.

File organization follow-up from #108 review: AX overrides,
AccessibilityText and CachedValue move to GhosttySurfaceAccessibility.swift;
AX tests move from PerformanceRegressionTests.swift to
AccessibilityTests.swift.
